refactor: consolidate UnifiedSearch into HybridSearch — one canonical search service (Issue #1542)
Merge UnifiedSearchService (616 lines) into HybridSearchService (495 lines),
keeping the best of both: RRF fusion and chunk search from HybridSearch,
query parsing, fuzzy title matching, tag search, and TUI convenience methods
from UnifiedSearch. Delete unified_search.py entirely.

- Add SearchQuery dataclass, parse_query(), _prepare_fts_query() with prefix
  wildcard expansion, fuzzy_search_titles(), _search_tags(), search_unified()
  async wrapper, get_recent_documents(), get_popular_tags(), get_document_by_id()
- Add FuzzyMatchDoc and PopularTagDict TypedDicts
- Add created_at and updated_at optional fields to HybridSearchResult
- Update palette_presenter.py and search_presenter.py to import from
  hybrid_search instead of unified_search
- Update docs/architecture.md to remove deleted file reference
